[REF] l10n_it_asset_management: Extract common tests setup
(cherry picked from commit d35fb61d8be2bad0e007097fe32b5f8d667bcaf7)
